requirements.txt
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
requests>=2.31.0
python-dotenv>=1.0.0
pydantic>=2.5.0
gradio>=4.0.0
smolagents>=0.1.0
duckduckgo-search>=6.1.0

opentelemetry-api>=1.21.0
opentelemetry-sdk>=1.21.0
opentelemetry-exporter-prometheus>=0.54b0
opentelemetry-instrumentation-fastapi>=0.46b0
opentelemetry-instrumentation-requests>=0.46b0
prometheus-client>=0.19.0